reorg(r5): move skills under core/skills (#486)
git mv skill_service, skill_importer, and skill_tools_bridge into
core/skills/. Leave explicit re-export shims at services/skill_* so
existing imports and patches keep working until R8.

Part of #486. No intentional behavior change.
