+++ core/skills/__init__.py
@@ -1,0 +1,5 @@
+"""Skills domain package (Reorg R5 / #486).
+
+Service and bridge modules for the Skill Builder live here. Callers should
+keep importing via the ``services.skill_*`` shims until Reorg R8 removes them.
+"""
